+++ src/openassetio-core/managerApi/HostSession.cpp
@@ -8,6 +8,10 @@ namespace openassetio {
 inline namespace OPENASSETIO_CORE_ABI_VERSION {
 namespace managerApi {
 
+HostSessionPtr HostSession::make(HostPtr host) {
+  return std::shared_ptr<HostSession>(new HostSession(std::move(host)));
+}
+
 HostSession::HostSession(HostPtr host) : host_{std::move(host)} {}
 
 HostPtr HostSession::host() const { return host_; }